requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.1
rembg==2.0.67
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
supabase==2.0.0
werkzeug==2.3.7
groq==0.4.1
pytrends==4.9.2
requests==2.31.0
